--- product-service/src/functions/hello/index.ts
@@ -1,19 +1,0 @@
-import schema from './schema';
-import { handlerPath } from '@libs/handler-resolver';
-
-export default {
-  handler: `${handlerPath(__dirname)}/handler.main`,
-  events: [
-    {
-      http: {
-        method: 'post',
-        path: 'hello',
-        request: {
-          schemas: {
-            'application/json': schema,
-          },
-        },
-      },
-    },
-  ],
-};
