--- product-service/src/functions/hello/schema.ts
@@ -1,7 +1,0 @@
-export default {
-  type: "object",
-  properties: {
-    name: { type: 'string' }
-  },
-  required: ['name']
-} as const;
