--- product-service/src/functions/index.ts
@@ -1,1 +1,0 @@
-export { default as hello } from './hello';
