+++ product-service/src/functions/products.ts
@@ -1,0 +1,23 @@
+export const getProductsList = {
+  handler: "src/handlers/getProductsList.handler",
+  events: [
+    {
+      http: {
+        method: "get",
+        path: "products/",
+      },
+    },
+  ],
+};
+
+export const getProductsById = {
+  handler: "src/handlers/getProductsById.handler",
+  events: [
+    {
+      http: {
+        method: "get",
+        path: "products/{id}",
+      },
+    },
+  ],
+};
