--- product-service/src/functions/products/schema.ts
@@ -1,7 +1,0 @@
-export default {
-  type: "object",
-  properties: {
-    id: { type: 'string' }
-  },
-  required: ['id']
-} as const;
