+++ product-service/src/handlers/getProductsById.ts
@@ -1,0 +1,14 @@
+import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
+import { formatJSONResponse } from "@libs/api-gateway";
+import { middyfy } from "@libs/lambda";
+import productService from "@services/index";
+
+export const handler = middyfy(
+  async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
+    const id = event.pathParameters.id;
+    const product = await productService.getProductsById(id);
+    return formatJSONResponse({
+      product,
+    });
+  }
+);
