--- product-service/src/libs/handler-resolver.ts
@@ -1,3 +1,0 @@
-export const handlerPath = (context: string) => {
-  return `${context.split(process.cwd())[1].substring(1).replace(/\\/g, '/')}`
-};
