+++ product-service/src/libs/lambda.ts
@@ -1,6 +1,11 @@
-import middy from "@middy/core"
-import middyJsonBodyParser from "@middy/http-json-body-parser"
+import middy from "@middy/core";
+import inputOutputLogger from "@middy/input-output-logger";
+import middyJsonBodyParser from "@middy/http-json-body-parser";
+import errorLogger from "@middy/error-logger";
 
 export const middyfy = (handler) => {
-  return middy(handler).use(middyJsonBodyParser())
-}
+  return middy(handler)
+    .use(inputOutputLogger())
+    .use(errorLogger())
+    .use(middyJsonBodyParser());
+};
