+++ product-service/src/mocks/data/products.ts
@@ -1,4 +1,4 @@
-import { Product } from "../models/Product";
+import { Product } from "@models/Product";
 
 export const products: Product[] = [
   {
