+++ product-service/src/models/Product.ts
